+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-## [project-title] Changelog
+## [Azure Functions Isolated Worker Samples](https://github.com/Azure-Samples/azure-functions-isloated-worker-sample) Changelog
 
+<!-- 
 <a name="x.y.z"></a>
 # x.y.z (yyyy-mm-dd)
 
@@ -10,4 +11,4 @@
 * ...
 
 *Breaking Changes*
-* ...
+* ... -->